[Refactor] Extract takeOwnership and takeOwnershipWithTime
