+++ src/systems/bit-media-frames.js
@@ -16,7 +16,7 @@ import { addObject3DComponent } from "../utils/jsx-entity";
 import { updateMaterials } from "../utils/material-utils";
 import { MediaType } from "../utils/media-utils";
 import { cloneObject3D, setMatrixWorld } from "../utils/three-utils";
-import { takeOwnership } from "../bit-systems/networking";
+import { takeOwnership } from "../utils/take-ownership";
 
 const EMPTY_COLOR = 0x6fc0fd;
 const HOVER_COLOR = 0x2f80ed;
